# docs/swagger.yaml
basePath: /
info:
  contact: {}
paths:
  /cocktail:
    put:
      consumes:
      - application/json
      description: Add IBA Cocktail
      parameters:
      - in: formData
        name: Garnish
        type: string
      - in: formData
        name: Id
        type: integer
      - in: formData
        name: Ingredients
        type: string
      - in: formData
        name: Method
        type: string
      - in: formData
        name: Name
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Add Cocktail
  /cocktail:id:
    delete:
      consumes:
      - application/json
      description: Delete One IBA Cocktail
      parameters:
      - description: Id cocktail
        in: query
        name: id
        required: true
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Delete Cocktail by Id
    get:
      consumes:
      - application/json
      description: get one IBA Cocktail
      parameters:
      - description: Id cocktail
        in: query
        name: id
        required: true
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: get Cocktail by Id
  /cocktails/:id:
    get:
      consumes:
      - application/json
      description: List IBA Cocktails
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: List Cocktails
  /login:
    post:
      consumes:
      - application/json
      description: Request Login
      parameters:
      - in: formData
        name: email
        type: string
      - in: formData
        name: password
        type: string
      produces:
      - application/json
      responses: {}
      summary: Login
securityDefinitions:
  "":
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"